stlying is really a tedious thing
You spend hours trying different fonts and styling, in the end you are
still not satisfied, as there seems to be no such a thing as the
ultimate or the best styling which is something you can usually get to
when doing programming. This may be the reason for which many great
people in the history despise the fashion industry.