haha... back to roboto after fiddling font for hours
the pair of go and go mono is a good one, but i'm now used to roboto
mono. it's hard and unnecessary for me to change font just for the
sake of it.

and, i say the typefoundry thing should die. in china, you don't see
many font types, just Song and Hei. it's so simple. the effort spent
on styling is unworthy and should be baned. even there are many
different types available for chinese fonts, like Li, Weibei, there
is almost no chinese websites using it for body text.